Fixed Db and Added ability to choose between Fairs
I updated the rssinfo table within the group's babbage and the .sql file
in order to allow for the correct data to be inserted into their
respective fields.

Additionally, I added the ability to switch between fairs. This loads
(via $_POST) a different XML file and therefore shows a different list
of companies once the User returns to the "Home" Page after he/she
selects which fair he wants to view.