Map Update
Committing map update.  There is now a toggle switch to control whether
or not you want to view the map with directions.  After hitting submit
now the inputs disappear and the text directions slide up for better
viewing.  There is also a search again button so that we can hide the
rest but still allow the user to search without reloading theirselves